+++ src/lib/features/feature-toggle/createFeatureToggleService.ts
@@ -14,7 +14,7 @@ import GroupStore from '../../db/group-store';
 import { AccountStore } from '../../db/account-store';
 import { AccessStore } from '../../db/access-store';
 import RoleStore from '../../db/role-store';
-import EnvironmentStore from '../project-environment/environment-store';
+import EnvironmentStore from '../project-environments/environment-store';
 import { Db } from '../../db/db';
 import { IUnleashConfig } from '../../types';
 import FakeEventStore from '../../../test/fixtures/fake-event-store';
@@ -28,7 +28,7 @@ import FakeGroupStore from '../../../test/fixtures/fake-group-store';
 import { FakeAccountStore } from '../../../test/fixtures/fake-account-store';
 import FakeAccessStore from '../../../test/fixtures/fake-access-store';
 import FakeRoleStore from '../../../test/fixtures/fake-role-store';
-import FakeEnvironmentStore from '../project-environment/fake-environment-store';
+import FakeEnvironmentStore from '../project-environments/fake-environment-store';
 import EventStore from '../../db/event-store';
 import {
     createChangeRequestAccessReadModel,
