+++ src/lib/services/index.ts
@@ -22,7 +22,7 @@ import SettingService from './setting-service';
 import SessionService from './session-service';
 import UserFeedbackService from './user-feedback-service';
 import FeatureToggleService from '../features/feature-toggle/feature-toggle-service';
-import EnvironmentService from '../features/project-environment/environment-service';
+import EnvironmentService from '../features/project-environments/environment-service';
 import FeatureTagService from './feature-tag-service';
 import ProjectHealthService from './project-health-service';
 import UserSplashService from './user-splash-service';
